Moved new JS code to a JS file
