+++ js/src/admin.js
@@ -1,0 +1,7 @@
+/**
+ * @package Polylang
+ */
+
+import ajaxFilter from './lib/ajax-filter.js';
+
+ajaxFilter( pll_admin.ajax_filter );
